+++ Sources/App/StationView.swift
@@ -16,16 +16,12 @@ import FairApp
 import FairKit
 import AVKit
 import AVFoundation
-import VideoToolbox
+//import VideoToolbox
 import WebKit
 import SwiftUI
 #if os(iOS)
 import MediaPlayer
 #endif
-#if canImport(TabularDataXXX)
-import TabularData
-#endif
-
 
 struct StationView: View {
     let station: Station
